# fix warnings more icc11 fixes - mostly type conversion and noncopyable (now made a macro)
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@6536 3db68df2-c116-0410-a063-a993310a9797

+++ source/lib/cache_adt.h
@@ -618,7 +618,7 @@ template<class Item, class Divider> struct CacheEntry
 	{
 		size = size_;
 		cost = cost_;
-		credit = cost;
+		credit = (float)cost;
 
 		// else divider will fail
 		debug_assert(size != 0);
